refactor: move bookmark stuff into its own folder
